* Defaulting file sender path to temp directory * Fixed bug in resolving variables in the sending path. `replaceAll` uses regex in the replacing value, which caused problems. * Removing logic to clean profiles from resources and adding explicit profile assertions. This is now being handled by the results of the CQF-Ruler (as part of the CQL logic). * Adding warning statement in logs when resources exist in the bundle without profile assertions, which would cause them not to be validated. * Moving `expandEnVars` to helper class so that it may be re-used by other classes in the future
